lxd/migrate: Do() push mode
Adapt Do() method of migrationSink struct to support push mode. Note that the
boolean we take to indicate whether we need live migration support or not is
always taken to be src.live and not dest.live. This is because src will always
hold information about the container that is actually moved. The same goes for
src.container which will always hold the actual container struct.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>