client: MigrateFrom() support push mode
In push mode the client will connect to the websockets from the source server
and do a post on the target server. The target server will be informed to expect
a push and in response return a set of secrets to the client. The client will
then use those secrets to connect to a set of websockets matching those of the
source server. The client will serve as a proxy between and source and target
server and relay the data accordingly.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>